requirements.txt
tqdm
numpy
utils
cython
Pillow
PyYAML
pandas
Jinja2
fastapi
uvicorn
ipykernel
matplotlib
opencv-python
torch-summary
dill==0.3.5.1
albumentations
python-multipart
email-validator>=2.0
#pycocotools
google-cloud-storage
apache-airflow
#==2.4.3
-e .